requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
common==0.1.2
config==0.5.0.post0
ddt==1.4.1
fastavro==0.23.4
grpcio==1.29.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.2
psycopg2-binary==2.8.5
pymmh3==0.0.5
pytz==2020.1
ratelimit==2.2.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
